changed method name to mark_board
