+++ src/board.py
@@ -24,7 +24,7 @@ class Board:
         return board[input_index]
 
     # go into a player class
-    def mark_board_with_user_selection(self, user_input, board, total_marks_on_board):
+    def mark_board(self, user_input, board, total_marks_on_board):
         self.assign_board_index_to_current_player_mark(
             user_input, board, total_marks_on_board
         )
