+++ test/test_game.py
@@ -1,6 +1,7 @@
 import unittest
 from unittest.mock import patch
 
+from src.board import Board
 from src.game import Game
 
 
@@ -75,3 +76,42 @@ class TestGame(unittest.TestCase):
         game = Game()
         output = game.process_user_input()
         self.assertEqual(output, ["1", "X", "3", "4", "5", "6", "7", "8", "9"])
+
+    def test_board_is_marked_with_user_selection_when_no_marks_and_X_is_going_first(
+        self,
+    ):
+        board = Board()
+        current_board = ["1", "2", "3", "4", "5", "6", "7", "8", "9"]
+        new_board_with_marks = ["X", "2", "3", "4", "5", "6", "7", "8", "9"]
+        user_input = 1
+        total_marks_on_board = 0
+        self.assertEqual(
+            new_board_with_marks,
+            board.mark_board(user_input, current_board, total_marks_on_board),
+        )
+
+    def test_board_is_marked_with_user_selection_when_one_mark_and_O_is_next_player(
+        self,
+    ):
+        board = Board()
+        current_board = ["X", "2", "3", "4", "5", "6", "7", "8", "9"]
+        new_board_with_marks = ["X", "O", "3", "4", "5", "6", "7", "8", "9"]
+        user_input = 2
+        total_marks_on_board = 1
+        self.assertEqual(
+            new_board_with_marks,
+            board.mark_board(user_input, current_board, total_marks_on_board),
+        )
+
+    def test_board_is_marked_with_user_selection_when_two_marks_and_X_is_next_player(
+        self,
+    ):
+        board = Board()
+        current_board = ["X", "O", "3", "4", "5", "6", "7", "8", "9"]
+        new_board_with_marks = ["X", "O", "X", "4", "5", "6", "7", "8", "9"]
+        user_input = 3
+        total_marks_on_board = 2
+        self.assertEqual(
+            new_board_with_marks,
+            board.mark_board(user_input, current_board, total_marks_on_board),
+        )
